backend/requirements.txt
Django==4.2.5
djangorestframework==3.15.0
django-cors-headers==3.14.0
joblib==1.3.2
scikit-learn==1.3.2
pandas==2.3.2
sentence-transformers
torch